script/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
protobuf==4.21.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.37
uritemplate==4.1.1
urllib3==1.26.9
xmltodict==0.13.0
